+++ src/main/java/com/adventofcode/y2019/intcode/values/ValueSource.java
@@ -1,5 +1,5 @@
 package com.adventofcode.y2019.intcode.values;
 
 public abstract class ValueSource {
-    public abstract int getValue();
+    public abstract long getValue();
 }
